requirements.txt
django==5.1.4
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
django-filter==24.3